LogsMate/LogsMate Backend/requirements.txt
langchain
langchain-community
sentence-transformers
langchain-groq
torch
chromadb
torchvision
torchaudio
torchtext
langchain-mistralai
mistralai==0.4.2
rank_bm25
tiktoken
FlagEmbedding
optuna
fpdf
tqdm
httpx
numpy
fastapi
uvicorn
python-multipart
langchain-google-genai